fix: add more words and alternative spellings
Some of the words were spelled in a way that is not really used a lot, for example nobody would write the word `dolboy'eb` in this way, it would probably be spelled `dolboeb`.

Also, the alternative spelling with the letter ё was added.

Some common words were lacking, like "suka" and "blyat".

Removed some obviously incorrect words, like "what's difference" (not a profanity).

Removed the word "gol" because it's not obscene? It means "goal" as in soccer?